+++ index.html
@@ -227,6 +227,6 @@
     </section>
   </template>
 
-  <script src="./js/main.js"></script>
+  <script src="./js/main.js" type="module"></script>
 </body>
 </html>
